Remove `tsl::robin_map` dependency
Since we now have Boost 1.81 readily available, we can make use of the
superior set and map type it contains and remove a dependency from VAST.
Swapping this out was pretty straightforward, and only required minor
adjusmtents.
