+++ Dockerfile
@@ -109,7 +109,6 @@ RUN apt-get update && \
       libyaml-cpp0.7 \
       lsb-release \
       openssl \
-      robin-map-dev \
       wget && \
     wget "https://apache.jfrog.io/artifactory/arrow/$(lsb_release --id --short | tr 'A-Z' 'a-z')/apache-arrow-apt-source-latest-$(lsb_release --codename --short).deb" && \
     apt-get -y --no-install-recommends install \
