+++ libvast/aux/README.md
@@ -10,9 +10,6 @@ Repositories
 ### submodules
 - [CAF](https://github.com/actor-framework/actor-framework) https://github.com/actor-framework/actor-framework
 
-### subtrees
-- [robin-map](https://github.com/Tessil/robin-map/): git@github.com:Tessil/robin-map.git
-
 Adding a New Repository
 -----------------------
