--- libvast/aux/robin-map/.codecov.yml
@@ -1,1 +1,0 @@
-comment: off
