+++ web/docs/setup/build.md
@@ -41,7 +41,6 @@ Every [release](https://github.com/tenzir/vast/releases) of VAST includes an
 |✓|[spdlog](https://github.com/gabime/spdlog)|>= 1.5|Required for logging.|
 |✓|[fmt](https://fmt.dev)|>= 8.1.1|Required for formatted text output.|
 |✓|[xxHash](https://github.com/Cyan4973/xxHash)|>= 0.8.0|Required for computing fast hash digests.|
-|✓|[robin-map](https://github.com/Tessil/robin-map)|>= 0.6.3|Fast hash map and hash set using robin hood hashing. (Bundled as subtree.)|
 |✓|[fast_float](https://github.com/FastFloat/fast_float)|>= 3.2.0|Required for parsing floating point numbers. (Bundled as submodule.)|
 ||[yarn](https://yarnpkg.com)||Required for building the web frontend.|
 ||[libpcap](https://www.tcpdump.org)||Required for PCAP import, export, and pivoting to and from PCAP traces.|
